requirements.txt
streamlit==1.39.0
openai>=1.0.0
chromadb==0.5.5
python-dotenv==1.0.1
tabulate==0.9.0
azure-ai-formrecognizer==3.3.3
azure-core==1.36.0
tiktoken==0.7.0
